removed some compiler warnings
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@737 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ rdr/ZlibOutStream.cxx
@@ -62,7 +62,7 @@ void ZlibOutStream::setUnderlying(OutStream* os)
 
 int ZlibOutStream::length()
 {
-  return offset + ptr - start;
+  return (int)(offset + ptr - start);
 }
 
 void ZlibOutStream::flush()
